Revert precision for X/YResolution to float for now, to avoid unnecessary changes in other programs.

+++ include/NFmiNetCDF.h
@@ -77,8 +77,8 @@ class NFmiNetCDF {
     bool FlipY();
     void FlipY(bool theYFlip);
  
-    double XResolution();
-    double YResolution();
+    float XResolution();
+    float YResolution();
 
     NcVar* GetVariable(const std::string& varName);
     bool CoordinatesInRowMajorOrder(const NcVar* var);
